+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Braintree iOS SDK Release Notes
 
-## unreleased
+## unreleased (5.0.0-beta2)
+* Add SPM support for `BraintreeDataCollector`
+* Bump Kount to v4.0.4.3 pre-release (provides an xcframework for SPM)
 * Fix Xcode 12.3 issue with building PPRiskMagnes.framework for iOS + iOS Simulator
 * Breaking Changes
   * Remove `type` and `nonce` params on `BTApplePayCardNone` initializer